+++ ui/appui-react/src/test/statusbar/StatusBar.test.tsx
@@ -4,40 +4,25 @@
  *--------------------------------------------------------------------------------------------*/
 import { expect } from "chai";
 import * as React from "react";
+import { render, screen } from "@testing-library/react";
+import { NoRenderApp } from "@itwin/core-frontend";
 import {
-  act,
-  fireEvent,
-  render,
-  screen,
-  waitForElementToBeRemoved,
-} from "@testing-library/react";
-import {
-  ActivityMessageDetails,
-  ActivityMessageEndReason,
-  NoRenderApp,
-  NotifyMessageDetails,
-  OutputMessagePriority,
-  OutputMessageType,
-} from "@itwin/core-frontend";
-import {
-  AppNotificationManager,
+  ConfigurableCreateInfo,
   MessageManager,
   StatusBar,
   StatusBarCenterSection,
   StatusBarLeftSection,
   StatusBarRightSection,
   StatusBarSpaceBetween,
+  StatusBarWidgetControl,
 } from "../../appui-react";
 import TestUtils from "../TestUtils";
 
 describe("StatusBar", () => {
-  let notifications: AppNotificationManager;
-
   before(async () => {
+    await NoRenderApp.startup();
     await TestUtils.initializeUiFramework();
 
-    notifications = new AppNotificationManager();
-    await NoRenderApp.startup();
     MessageManager.clearMessages();
   });
 
@@ -45,240 +30,26 @@ describe("StatusBar", () => {
     TestUtils.terminateUiFramework();
   });
 
-  it("StatusBar should render a Toast message", async () => {
-    render(<StatusBar />);
+  it("StatusBar should render children correctly", () => {
+    render(<StatusBar>Hello</StatusBar>);
 
-    const details = new NotifyMessageDetails(
-      OutputMessagePriority.None,
-      "A brief message.",
-      "A detailed message."
-    );
-    act(() => {
-      notifications.outputMessage(details);
-    });
-    expect(await screen.findByText("A brief message.")).to.be.not.null;
-
-    act(() => {
-      MessageManager.closeAllMessages();
-    });
-    await waitForElementToBeRemoved(screen.queryByText("A brief message."));
+    expect(screen.getByText("Hello")).to.be.not.null;
   });
 
-  it("StatusBar should render a Sticky message", async () => {
-    render(<StatusBar />);
-
-    const details = new NotifyMessageDetails(
-      OutputMessagePriority.Info,
-      "A brief message.",
-      "A detailed message.",
-      OutputMessageType.Sticky
+  it("StatusBar should render widgetControl correctly", () => {
+    render(
+      <StatusBar
+        widgetControl={
+          new (class extends StatusBarWidgetControl {
+            public getReactNode(): React.ReactNode {
+              return "ReactNodeContent";
+            }
+          })(new ConfigurableCreateInfo("a", "b", "c"), {})
+        }
+      />
     );
-    act(() => {
-      notifications.outputMessage(details);
-    });
-    expect(await screen.findByText("A brief message.")).to.be.not.null;
 
-    act(() => {
-      MessageManager.closeAllMessages();
-    });
-    await waitForElementToBeRemoved(screen.queryByText("A brief message."));
-  });
-
-  it("Sticky message should close on button click", async () => {
-    render(<StatusBar />);
-
-    const details = new NotifyMessageDetails(
-      OutputMessagePriority.Error,
-      "A brief message.",
-      "A detailed message.",
-      OutputMessageType.Sticky
-    );
-    act(() => {
-      notifications.outputMessage(details);
-    });
-    expect(await screen.findByText("A brief message.")).to.be.not.null;
-
-    const closeButton = screen.getByRole("button", { name: "Close" });
-    fireEvent.click(closeButton);
-    await waitForElementToBeRemoved(screen.queryByText("A brief message."));
-    act(() => {
-      MessageManager.clearMessages();
-    });
-  });
-
-  it("StatusBar should render an Activity message", async () => {
-    render(<StatusBar />);
-
-    const details = new ActivityMessageDetails(true, true, false);
-    notifications.setupActivityMessage(details);
-    act(() => {
-      notifications.outputActivityMessage("Message text", 50);
-    });
-    expect(await screen.findByText("Message text")).to.be.not.null;
-
-    act(() => {
-      notifications.endActivityMessage(ActivityMessageEndReason.Completed);
-    });
-    await waitForElementToBeRemoved(screen.queryByText("Message text"));
-  });
-
-  it("Activity message should be canceled", async () => {
-    render(<StatusBar />);
-
-    const details = new ActivityMessageDetails(true, true, true);
-    notifications.setupActivityMessage(details);
-    act(() => {
-      notifications.outputActivityMessage("Message text", 50);
-    });
-    expect(await screen.findByText("Message text")).to.be.not.null;
-
-    const cancelLink = await screen.findByText("dialog.cancel");
-    fireEvent.click(cancelLink);
-
-    await waitForElementToBeRemoved(screen.queryByText("Message text"));
-  });
-
-  it("Activity message should be dismissed", async () => {
-    render(<StatusBar />);
-
-    const details = new ActivityMessageDetails(true, true, true);
-    notifications.setupActivityMessage(details);
-    notifications.setupActivityMessage(details);
-    act(() => {
-      notifications.outputActivityMessage("Message text", 50);
-    });
-    expect(await screen.findByText("Message text")).to.be.not.null;
-
-    const closeButton = screen.getByRole("button", { name: "Close" });
-    fireEvent.click(closeButton);
-    await waitForElementToBeRemoved(screen.queryByText("Message text"));
-  });
-
-  it("StatusBar should render Toast, Sticky & Activity messages", async () => {
-    render(<StatusBar />);
-
-    const details1 = new NotifyMessageDetails(
-      OutputMessagePriority.Warning,
-      "A brief message.",
-      "A detailed message."
-    );
-    const details2 = new NotifyMessageDetails(
-      OutputMessagePriority.None,
-      "A brief sticky message.",
-      "A detailed message.",
-      OutputMessageType.Sticky
-    );
-    const details3 = new ActivityMessageDetails(true, true, true);
-    notifications.setupActivityMessage(details3);
-    act(() => {
-      notifications.outputMessage(details1);
-      notifications.outputMessage(details2);
-      notifications.outputActivityMessage("Message text", 50);
-    });
-    expect(await screen.findByText("A brief message.")).to.be.not.null;
-    expect(await screen.findByText("A brief sticky message.")).to.be.not.null;
-    expect(await screen.findByText("Message text")).to.be.not.null;
-    act(() => {
-      notifications.endActivityMessage(ActivityMessageEndReason.Completed);
-    });
-    act(() => {
-      MessageManager.closeAllMessages();
-    });
-    await waitForElementToBeRemoved(
-      screen.queryByText("A brief sticky message.")
-    );
-  });
-
-  it("StatusBar should render maximum of 3 Sticky messages", async () => {
-    MessageManager.maxDisplayedStickyMessages = 3;
-
-    render(<StatusBar />);
-
-    const details1 = new NotifyMessageDetails(
-      OutputMessagePriority.None,
-      "A brief message 1.",
-      undefined,
-      OutputMessageType.Sticky
-    );
-    act(() => {
-      notifications.outputMessage(details1);
-    });
-    expect(await screen.findByText("A brief message 1.")).to.be.not.null;
-    const details2 = new NotifyMessageDetails(
-      OutputMessagePriority.None,
-      "A brief message 2.",
-      undefined,
-      OutputMessageType.Sticky
-    );
-    act(() => {
-      notifications.outputMessage(details2);
-    });
-    expect(await screen.findByText("A brief message 2.")).to.be.not.null;
-    const details3 = new NotifyMessageDetails(
-      OutputMessagePriority.None,
-      "A brief message 3.",
-      undefined,
-      OutputMessageType.Sticky
-    );
-    act(() => {
-      notifications.outputMessage(details3);
-    });
-    expect(await screen.findByText("A brief message 3.")).to.be.not.null;
-
-    const details4 = new NotifyMessageDetails(
-      OutputMessagePriority.None,
-      "A brief message 4.",
-      undefined,
-      OutputMessageType.Sticky
-    );
-    act(() => {
-      notifications.outputMessage(details4);
-    });
-    await waitForElementToBeRemoved(screen.queryByText("A brief message 1."));
-    expect(await screen.findByText("A brief message 4.")).to.be.not.null;
-
-    act(() => {
-      MessageManager.closeAllMessages();
-    });
-    await waitForElementToBeRemoved(screen.queryByText("A brief message 4."));
-  });
-
-  it("StatusBar should not render a Pointer message", () => {
-    render(<StatusBar />);
-
-    const details = new NotifyMessageDetails(
-      OutputMessagePriority.Info,
-      "A brief message.",
-      "A detailed message.",
-      OutputMessageType.Pointer
-    );
-    act(() => {
-      notifications.outputMessage(details);
-    });
-
-    expect(screen.queryByText("A brief message.")).to.be.null;
-  });
-
-  it("StatusBar should clear messages", async () => {
-    render(<StatusBar />);
-
-    const details = new NotifyMessageDetails(
-      OutputMessagePriority.Info,
-      "A brief toast message.",
-      "A detailed message.",
-      OutputMessageType.Sticky
-    );
-    act(() => {
-      notifications.outputMessage(details);
-    });
-    expect(await screen.findByText("A brief toast message.")).to.be.not.null;
-
-    act(() => {
-      MessageManager.clearMessages();
-    });
-    await waitForElementToBeRemoved(
-      screen.queryByText("A brief toast message.")
-    );
+    expect(screen.getByText("ReactNodeContent")).to.be.not.null;
   });
 
   it("StatusBarSpaceBetween should render correctly", () => {
